requirements_common.txt
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.8.6
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.2.0
analytics-python==1.4.post1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
av==11.0.0
azure-cognitiveservices-speech==1.35.0
backcall==0.2.0
backoff==1.10.0
baidu-aip==4.16.11
bardapi==0.1.33
bce-python-sdk==0.8.97
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.22.1
bilibili-api==4.1.0
bilibili-api-dev==16.2.0b1
bilibili-api-python==16.1.0
blinker==1.7.0
blis==0.7.9
Brotli==1.1.0
brotlipy==0.7.0
browser-cookie3==0.19.1
cachetools==5.3.1
catalogue==2.0.8
chardet==4.0.0
ChatterBot==1.1.0a7
click==8.1.3
coloredlogs==15.0.1
confection==0.1.0
contourpy==1.2.0
cssutils==2.7.1
ctranslate2==4.0.0
curl-cffi==0.5.7
cycler==0.12.1
cymem==2.0.7
dataclasses==0.6
dataclasses-json==0.5.9
decorator==5.1.1
deep-translator==1.11.4
distro==1.8.0
EasyProcess==1.1
edge-tts==6.1.6
elevenlabs==0.2.8
emoji==2.9.0
entrypoint2==1.1
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.103.2
fastapi-socketio==0.0.10
faster-whisper==1.0.0
ffmpy==0.3.1
filelock==3.12.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.8.2
google-generativeai==0.3.1
googleapis-common-protos==1.56.4
gptcache==0.1.34
gradio==3.0
gradio_client==0.7.3
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.0
httpx==0.25.2
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
ifaddr==0.2.0
importlib-resources==6.1.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyboard==0.13.5
kiwisolver==1.4.5
langchain==0.0.142
langcodes==3.3.0
langdetect==1.0.9
langid==1.1.6
linkify-it-py==2.0.3
loguru==0.7.0
lxml==4.9.3
lz4==4.3.2
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mashumaro==3.12
mathparse==0.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nicegui==1.4.5
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
onnxruntime==1.16.3
openai==1.6.1
openapi-schema-pydantic==1.2.4
orjson==3.9.9
pandas==2.2.0
paramiko==3.4.0
parso==0.8.3
pathy==0.10.2
pickleshare==0.7.5
Pillow==10.1.0
playwright==1.36.0
preshed==3.0.8
profanity==1.1
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==3.20.3
protobuf-inspector==0.2
protobuf3-to-dict==0.1.5
pscript==0.7.7
pure-eval==0.2.2
pure-protobuf==3.0.0a5
pyahocorasick==2.0.0
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycryptodome==3.19.0
pycryptodomex==3.19.1
pydantic==1.10.13
pydantic-settings==2.0.0
pydantic_core==2.16.1
pydub==0.25.1
pyee==9.0.4
pygame==2.5.0
PyGetWindow==0.0.9
Pygments==2.15.1
pygtrans==1.5.3
pyhandytools==0.0.7
PyJWT==2.8.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.1.1
pypdf==3.11.1
PyPDF2==3.0.1
pyperclip==1.8.2
pypinyin==0.49.0
pypng==0.20220715.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
pyreadline3==3.4.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytchat==0.5.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-multipart==0.0.6
python-socketio==5.10.0
pytweening==1.0.7
pytz==2023.3
pyvirtualcam==0.10.2
PyYAML==6.0
qrcode==7.4.2
qrcode-terminal==0.8
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
referencing==0.33.0
regex==2023.6.3
requests==2.31.0
revTongYi==0.0.5.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruff==0.2.0
safetensors==0.3.1
schedule==1.2.0
scikit-learn==1.3.0
scipy==1.11.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
simple-websocket==1.0.0
slack-sdk==3.21.3
smart-open==6.3.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sparkdesk-api==1.5.0
SpeechRecognition==3.10.0
SQLAlchemy==1.3.24
srsly==2.4.6
sseclient-py==1.8.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.4.0
TikTokLive==5.0.8
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
vbuild==0.8.2
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.6
websocket-client==1.6.3
websocket-client-py3==0.15.0
websockets==11.0.3
webuiapi==0.9.3
Werkzeug==3.0.1
win32-setctime==1.1.0
wincertstore==0.2
wsproto==1.2.0
xingchen==1.0.7
yarl==1.9.4
zhipuai==1.0.7
zstandard==0.19.0
git+https://gitee.com/ikaros-521/WenxinWorkshop-Python-SDK
git+https://gitee.com/ikaros-521/blivedm
dashscope==1.14.1
pygetwindow
opencv-python==4.9.0.80